class EduHub integrated into metaclass api
